# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TKOH OMS Backend Server",
        "title": "TKOH OMS",
        "contact": {
            "name": "So Nim Wai"
        },
        "license": {
            "name": "LSCM"
        },
        "version": "0.1"
    },
    "host": "20.239.95.146",
    "basePath": "/oms/",
    "paths": {
...
        "/reportJobStatus": {
            "post": {
                "description": "Receive the delivery job updated status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Management"
                ],
                "summary": "Report Job Status.",
                "parameters": [
                    {
                        "description": "Return Job Status Parameters",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReportJobStatusDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseHeader"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
        "/reportSystemStatus": {
            "post": {
                "description": "Get current system status.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Management"
                ],
                "summary": "Report System Status.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderManagement.ReportSystemStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
...
        "/triggerHandlingOrder": {
            "get": {
                "description": "Notify system users are ready to handle the current order.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Management"
                ],
                "summary": "Trigger Delivery Order.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order IDs",
                        "name": "orderIds",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Schedule IDs",
                        "name": "scheduleId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderManagement.TriggerHandlingOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ReportJobStatusDTO": {
            "type": "object",
            "properties": {
                "est": {
                    "type": "string"
                },
                "eta": {
                    "type": "string"
                },
                "jobId": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "messageTime": {
                    "type": "string"
                },
                "processingStatus": {
                    "type": "string"
                },
                "robotId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
        "models.FailResponse": {
            "type": "object",
            "properties": {
                "header": {
                    "$ref": "#/definitions/models.FailResponseHeader"
                }
            }
        },
        "models.FailResponseHeader": {
            "type": "object",
            "properties": {
                "failedReason": {
                    "type": "string"
                },
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "models.ResponseHeader": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
...
        "orderManagement.OrderListBody": {
            "type": "object",
            "properties": {
                "orderList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "actualArrivalTime": {
                                "type": "string"
                            },
                            "endLocationId": {
                                "type": "integer"
                            },
                            "endLocationName": {
                                "type": "string"
                            },
                            "endTime": {
                                "type": "string"
                            },
                            "expectingArrivalTime": {
                                "type": "string"
                            },
                            "expectingDeliveryTime": {
                                "type": "string"
                            },
                            "expectingStartTime": {
                                "type": "string"
                            },
                            "failedReason": {
                                "type": "string"
                            },
                            "orderCreatedBy": {
                                "type": "integer"
                            },
                            "orderCreatedType": {
                                "type": "string"
                            },
                            "orderId": {
                                "type": "integer"
                            },
                            "orderStatus": {
                                "type": "string"
                            },
                            "orderType": {
                                "type": "string"
                            },
                            "processingStatus": {
                                "type": "string"
                            },
                            "scheduleId": {
                                "type": "integer"
                            },
                            "startLocationId": {
                                "type": "integer"
                            },
                            "startLocationName": {
                                "type": "string"
                            },
                            "startTime": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "orderManagement.ReportSystemStatusResponse": {
            "type": "object",
            "properties": {
                "systemState": {
                    "type": "string"
                },
                "systemStatus": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "orderManagement.TriggerHandlingOrderResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/orderManagement.OrderListBody"
                },
                "header": {
                    "$ref": "#/definitions/models.ResponseHeader"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}